+++ src/test/java/it/gov/pagopa/payhub/activities/activity/paymentsreporting/PaymentsReportingIngestionFlowFileActivityImplTest.java
@@ -1,7 +1,7 @@
 package it.gov.pagopa.payhub.activities.activity.paymentsreporting;
 
 import it.gov.pagopa.payhub.activities.dao.IngestionFlowFileDao;
-import it.gov.pagopa.payhub.activities.dto.paymentsreporting.IngestionFlowFileDTO;
+import it.gov.pagopa.payhub.activities.dto.IngestionFlowFileDTO;
 import it.gov.pagopa.payhub.activities.dto.reportingflow.PaymentsReportingIngestionFlowFileActivityResult;
 import it.gov.pagopa.payhub.activities.service.ingestionflow.IngestionFlowFileRetrieverService;
 import org.junit.jupiter.api.BeforeEach;
@@ -38,7 +38,7 @@ class PaymentsReportingIngestionFlowFileActivityImplTest {
 			.ingestionFlowFileId(ingestionFlowFileId)
 			.fileName("valid-file.zip")
 			.filePathName("/valid/path")
-			.requestTokenCode("valid-token")
+			.codRequestToken("valid-token")
 			.build();
 
 		when(ingestionFlowFileDao.findById(ingestionFlowFileId)).thenReturn(Optional.of(mockFlowDTO));
@@ -77,7 +77,7 @@ class PaymentsReportingIngestionFlowFileActivityImplTest {
 			.ingestionFlowFileId(ingestionFlowFileId)
 			.fileName("valid-file.zip")
 			.filePathName("/valid/path")
-			.requestTokenCode("valid-token")
+			.codRequestToken("valid-token")
 			.build();
 
 		when(ingestionFlowFileDao.findById(ingestionFlowFileId)).thenReturn(Optional.of(mockFlowDTO));
